Implement download command.
