+++ cli/jdl.js
@@ -18,11 +18,10 @@
  */
 const chalk = require('chalk');
 const fs = require('fs');
-const https = require('https');
 const path = require('path');
 const cliUtils = require('./utils');
 const importJdl = require('./import-jdl');
-const packageJson = require('../package.json');
+const download = require('./download');
 
 const { logger } = cliUtils;
 
@@ -36,27 +35,6 @@ const toJdlFile = file => {
     return file;
 };
 
-const downloadFile = (url, filename) => {
-    return new Promise((resolve, reject) => {
-        logger.info(`Downloading file: ${url}`);
-        https
-            .get(url, response => {
-                if (response.statusCode !== 200) {
-                    return reject(new Error(`Error downloading ${url}: ${response.statusCode} - ${response.statusMessage}`));
-                }
-
-                logger.debug(`Creating file: ${path.join(filename)}`);
-                const fileStream = fs.createWriteStream(`${filename}`);
-                fileStream.on('finish', () => fileStream.close());
-                fileStream.on('close', () => resolve(filename));
-                response.pipe(fileStream);
-                return undefined;
-            })
-            .on('error', e => {
-                reject(e);
-            });
-    });
-};
 /**
  * JDL command
  * @param {string[][]} args arguments passed for import-jdl
@@ -76,25 +54,9 @@ module.exports = ([jdlFiles = []], options = {}, env, forkProcess) => {
     }
     const promises = jdlFiles.map(toJdlFile).map(filename => {
         if (!fs.existsSync(filename)) {
-            let url;
-            try {
-                const urlObject = new URL(filename);
-                url = filename;
-                filename = path.basename(urlObject.pathname);
-            } catch (_error) {
-                if (options.skipSampleRepository) {
-                    return Promise.reject(new Error(`Could not find ${filename}, make sure the path is correct.`));
-                }
-                url = new URL(filename, `https://raw.githubusercontent.com/jhipster/jdl-samples/v${packageJson.version}/`).toString();
-                filename = path.basename(filename);
-            }
-            return downloadFile(url, filename).catch(error => {
-                logger.info(error.message);
-                url = new URL(filename, 'https://raw.githubusercontent.com/jhipster/jdl-samples/main/').toString();
-                return downloadFile(url, filename);
-            });
+            return download([[filename]], options);
         }
         return Promise.resolve(filename);
     });
-    return Promise.all(promises).then(jdlFiles => importJdl(jdlFiles, options, env, forkProcess));
+    return Promise.all(promises).then(jdlFiles => importJdl(jdlFiles.flat(), options, env, forkProcess));
 };
